Registration & Login (10 files)
Registration page for Caretaker and Login page for both Caretaker and Patient
MainActivity.java(activity_main.xml) is the login page for caretaker. MyActivity.java(activity_main_my.xml) is the registration page for caretaker. patientActivity.java(patient_activity.xml) is the login page for patient. homepage(homepage.xml) is the main menu page and BackgroundTask.java is for connecting to MySQL database.